Memo

+++ app/Http/Controllers/BackupController.php
@@ -8,6 +8,7 @@ use App\Models\Undangan;
 use App\Models\Risalah;
 use App\Models\Divisi;
 use App\Models\Backup_Document;
+use App\Models\Kirim_Document;
 use Illuminate\Support\Facades\Auth;
 
 
@@ -113,10 +114,14 @@ class BackupController extends Controller
         $memo = Memo::withTrashed()
             ->where('id_memo', $id)
             ->first();
+        $kirim_documents = Kirim_Document::withTrashed()->where('id_document', $id)->where('jenis_document','memo')->get();
         if ($memo) {
             $memo->restore();
+            foreach ($kirim_documents as $kirim_memo) {
+                $kirim_memo->restore();
+            }
         } else {
-            dd($memo);
+           return redirect()->route('memo.backup')->with('failure', 'Memo tidak ditemukan.');
         }
         return redirect()->route('memo.backup')->with('success', 'Memo terpilih berhasil dipulihkan.');
     }
@@ -126,8 +131,15 @@ class BackupController extends Controller
         $undangan = Undangan::withTrashed()
             ->where('id_undangan', $id)
             ->first();
+        $kirim_documents = Kirim_Document::withTrashed()
+        ->where('id_document', $id)
+        ->where('jenis_document','undangan')
+        ->get();
         if ($undangan) {
             $undangan->restore();
+            foreach ($kirim_documents as $kirim_undangan) {
+                $kirim_undangan->restore();
+            }
         } else {
             dd($undangan);
         }
@@ -190,8 +202,18 @@ class BackupController extends Controller
     public function forceDelete($id)
     {
         $undangan = Undangan::withTrashed()->findOrFail($id);
-        $undangan->forceDelete();
-
+        $kirim_documents = Kirim_Document::withTrashed()
+        ->where('id_document', $id)
+        ->where('jenis_document','undangan')
+        ->get();
+        if ($undangan) {
+            $undangan->forceDelete();
+            foreach ($kirim_documents as $kirim_undangan) {
+                $kirim_undangan->forceDelete();
+            }
+        } else {
+            return redirect()->route('undangan.backup')->with('failure', 'Undangan tidak ditemukan.');
+        }
         return redirect()->route('undangan.backup')->with('success', 'Undangan berhasil dihapus permanen.');
     }
   
@@ -200,8 +222,15 @@ class BackupController extends Controller
      {
     
         $memo = Memo::onlyTrashed()->findOrFail($id);
-        $memo->forceDelete();
-
+        $kirim_document = Kirim_Document::onlyTrashed()->where('id_document', $id)->where('jenis_document','memo')->get();
+        if ($memo) {
+            $memo->forceDelete();
+            foreach ($kirim_document as $kirim_memo) {
+                $kirim_memo->forceDelete();
+            }
+        } else {
+           return redirect()->route('memo.backup')->with('failure', 'Memo tidak ditemukan.');
+        }
          return redirect()->route('memo.backup')->with('success', 'Memo terpilih berhasil dihapus permanen.');
      }
